# Add `StringNormalization` support to arrays and collections to fix Safari checksum issues