--- a/tests/Browser/StringNormalization/Component.php
+++ b/tests/Browser/StringNormalization/Component.php
@@ -10,6 +10,15 @@ class Component extends BaseComponent
     public $string = 'â';
     public $number = 0;
     public $array = ['â'];
+    public $recursiveArray = ['â', ['â']];
+    public $collection;
+    public $recursiveCollection;
+
+    public function mount()
+    {
+        $this->collection = collect(['â']);
+        $this->recursiveCollection = collect(['â', ['â']]);
+    }
 
     public function render()
     {