--- a/tests/Browser/StringNormalization/Model.php
+++ b/tests/Browser/StringNormalization/Model.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Tests\Browser\StringNormalization;
+
+use Illuminate\Database\Eloquent\Model as BaseModel;
+use Sushi\Sushi;
+
+class Model extends BaseModel
+{
+    use Sushi;
+
+    protected $rows = [
+        [
+            'id' => 1,
+            'name' => 'â'
+        ]
+    ];
+}