--- a/tests/Browser/StringNormalization/view.blade.php
+++ b/tests/Browser/StringNormalization/view.blade.php
@@ -1,5 +1,6 @@
 <div>
     <div>{{ $string }}</div>
     <div>{{ $number }}</div>
+    <div>{{ $array[0] }}</div>
     <button id="add_number" wire:click="addNumber">Add Number</button>
 </div>