--- a/tests/Browser/StringNormalization/view.blade.php
+++ b/tests/Browser/StringNormalization/view.blade.php
@@ -1,6 +1,9 @@
 <div>
     <div>{{ $string }}</div>
     <div>{{ $number }}</div>
-    <div>{{ $array[0] }}</div>
+    <pre>{{ print_r($array, true) }}</pre>
+    <pre>{{ print_r($recursiveArray, true) }}</pre>
+    <pre>{{ print_r($collection->toArray(), true) }}</pre>
+    <pre>{{ print_r($recursiveCollection->toArray(), true) }}</pre>
     <button id="add_number" wire:click="addNumber">Add Number</button>
 </div>