--- a/tests/Browser/TestCase.php
+++ b/tests/Browser/TestCase.php
@@ -32,7 +32,7 @@ class TestCase extends BaseTestCase
 {
     use SupportsSafari;
 
-    public static $useSafari = false;
+    public static $useSafari = true;
     public static $useAlpineV3 = false;
 
     public function setUp(): void