Update `StringNormalization` test to prove that special chars are not normalized and break Safari checksum.  
https://github.com/livewire/livewire/pull/4942 don't fix all cases.