@joshhanley 2022-12-08
@Balsakup thanks for the PR! All looks good for collections/ arrays/ strings.

What about the scenario where there is an eloquent model (or eloquent collection) which has special chars in attributes, does that break too?