@Balsakup 2022-12-09
@joshhanley, I added some tests for the Eloquent collections and models, and everything is working as expected