@joshhanley 2022-12-20
@Balsakup thanks! I've been running some performance checks on this to make sure we don't increase the processing time too much. Just waiting on Caleb to have a look too before actioning further 🙂