@joshhanley 2022-12-21
So we've run some benchmarks on this PR in comparison to the current master branch, and the results are as follows:

As this PR only affects collections and arrays, so that is what has been tested.

Benchmark 1 - array with 10,000 items and also a collection of that array - running 100 iterations of dehydrate method:

Master:
Total time: 3.125s
Average: 31.25ms

This PR:
Total time: 3.848s
Average: 38.48ms


Benchmark 2 - array with 10 nested arrays of 10,000 items (100,000 items total) and also a collection of that array - running 100 iterations of dehydrate method:

Master:
Total time: 2.692s
Average: 26.92s

This PR:
Total time: 12.956s
Average: 129.56ms

As you can see in the benchmark 1, running a combined total of 20,000 items between the array and the collection, the difference between master and this PR is nominal.

There is a noticeable difference in benchmark 2, between master and this PR, approx 5 times slower. But, that is essentially testing a combined total of 200,000 items between the array and the collection and it's still only ~130ms.



Our conclusion is, that those times are not ideal, but that is likely due to the recursion through the arrays/collections. It's also unlikely that someone will have that much data loaded into a single array.

As such, we are going to merge this, to ensure Safari is properly supported and doesn't throw the checksum issue.

There might be a better place to do this processing, like somewhere where we are already traversing the property data, which could make this more efficient. That's something we can look into in the future.